+++ client/ui/Results.jsx
@@ -10,15 +10,40 @@ import Paging from './Paging.jsx';
 
 // App component - represents the whole app
 class Results extends Component {
+
+componentDidMount() {
+  window.autosearch=false
+  window.searchedQuery=""
+  window.layersMessage=""
+}
+
+componentWillMount() {
+  window.layersMessage=""
+}
+
+componentDidUpdate () {
+  var r=this.props.results
+  window.autosearch=window.query==window.searchedQuery?false:true
+  // $(window.inputId)[0].value = window.query
+  if (window.autosearch && r && r.results.hits && r.results.hits.hits && r.results.hits.total.value==0) {
+      window.searchedQuery=window.query
+      // window.layersMessage=" all layers (except Chinese)"
+      this.globalSearch(1, this.props.limit)
+      // console.log("should have called")
+  } else {
+    window.autosearch=true;
+  }
+}
+
  render() {
    var r=this.props.results
-   //console.log(r);
    return <div className="Result base">
        {
         r?
            r.results.hits?
              (r.results.hits.hits && r.results.hits.total.value>0)?
              <div>
+              {window.layersMessage=""}
                <div className='resultCount' dir='ltr'><small>{r.results.hits.total.value} verses found ({r.results.took}ms).</small></div>
                {Object.keys(r.results.hits.hits).map((v) => (
                   <Verse key={r.results.hits.hits[v]._id}
@@ -28,7 +53,9 @@ class Results extends Component {
                       setVerse={this.setVerse.bind(this)}
                       options={this.props.options}
                       search={this.props.search.bind(this)}
-                      analyzers={this.props.analyzers}/>
+                      hideUnmatched={this.props.hideUnmatched}
+                      analyzers={this.props.analyzers}
+                      switchLayers={this.props.switchLayers}/>
 
                ))}
                <Paging total={r.results.hits.total.value}
@@ -37,8 +64,7 @@ class Results extends Component {
                   search={this.props.search.bind(this)}
                   page={this.props.page}
                   limit={this.props.limit} />
-              </div>:<div className="NoResult"><h4>Sorry, no results found</h4><hr/><h5>If like to perform a global search instead, click
-                      <a onClick={this.globalSearch.bind(this, 1, this.props.limit)}> here</a><br/><br/>See advance filter examples below for more details</h5><hr/><Help/></div>:"...":<Help/>
+              </div>:<div className="NoResult"><h4>Sorry, no results found {window.layersMessage}</h4><hr/><h5><br/>See advance filter examples below for more details</h5><hr/><Help/></div>:"...":<Help/>
        }
        </div>
  }
@@ -49,12 +75,28 @@ class Results extends Component {
  setPage(page,limit) {
     this.props.setPage(page,limit);
  }
-
+//{this.globalSearch(this, 1, this.props.limit)}
  globalSearch(page,limit) {
-   window.query = ":\""+window.query.replace(/"|\:/g,'') + "\""
-   $(window.inputId)[0].value = window.query
-   // this.props.options.map(y=>{y.state=true})
-   this.props.search(window.query,this.props.options,page,limit)
+   // console.log(page,limit);
+   // window.query = ":\""+window.query.replace(/"|\:/g,'') + "\""
+
+   // window.layersMessage+=" Searching in all layers "
+   this.props.switchLayers(true)
+   this.props.options.map(y=>{if (!y.id.match(/Chinese/i)){y.state=true}})
+
+   const myTimeout = setTimeout(function(){
+     // window.layersMessage+=" (done)"
+      $('button.Search').trigger("click")
+   },1000)
+
+
+   try {
+     // setTimeout(this.props.search, 1000, window.query, this.props.options,page,limit);
+     // this.props.search(window.query,this.props.options,page,limit)
+
+   } catch (e) {
+
+   }
 
    // alert(1)
  }
@@ -62,7 +104,7 @@ class Results extends Component {
 }
 
 Results.propTypes = {
- //results: PropTypes.object.isRequired
+ // results: PropTypes.object.isRequired
 }
 
 export default createContainer(props => {